+++ preprocess/pre-process-module-files.go
@@ -1,225 +1,11 @@
-package cli
+package preprocess
 
 import (
 	"fmt"
-	"github.com/gruntwork-io/terragrunt/errors"
 	"github.com/gruntwork-io/terragrunt/graph"
 	"github.com/gruntwork-io/terragrunt/options"
-	"github.com/gruntwork-io/terragrunt/util"
-	"github.com/hashicorp/hcl/v2"
-	"github.com/hashicorp/hcl/v2/hclsyntax"
-	"github.com/zclconf/go-cty/cty"
-	"os"
-	"path/filepath"
-	"strings"
+	"github.com/hashicorp/hcl/v2/hclwrite"
 )
-import "github.com/hashicorp/hcl/v2/hclwrite"
-
-const processHelp = `
-   Usage: terragrunt process <OUTPUT_DIR> [OPTIONS]
-
-   Description:
-     Pre-process the Terraform code in the current working directory and write the results to OUTPUT_DIR.
-   
-   Arguments:
-     OUTPUT_DIR: The directory where to write the pre-processed results.
-
-   Options:
-     TODO
-`
-
-const envsDirName = "envs"
-
-func runProcess(terragruntOptions *options.TerragruntOptions) error {
-	// First arg should be "process"; second should be output dir
-	if len(terragruntOptions.TerraformCliArgs) != 2 {
-		return fmt.Errorf("Unexpected number of arguments. Usage: %s", processHelp)
-	}
-
-	outputDir := terragruntOptions.TerraformCliArgs[1]
-
-	envNames, err := getEnvNames(terragruntOptions)
-	if err != nil {
-		return err
-	}
-
-	dependencyGraph, err := buildDependencyGraph(terragruntOptions)
-	if err != nil {
-		return err
-	}
-
-	if len(envNames) > 0 {
-		for _, envName := range envNames {
-			if err := createEnv(outputDir, &envName, dependencyGraph, terragruntOptions); err != nil {
-				return err
-			}
-		}
-
-		return nil
-	} else {
-		return createEnv(outputDir, nil, dependencyGraph, terragruntOptions)
-	}
-}
-
-func getEnvNames(terragruntOptions *options.TerragruntOptions) ([]string, error) {
-	envNames := []string{}
-
-	envsDir := filepath.Join(terragruntOptions.WorkingDir, envsDirName)
-	if !util.IsDir(envsDir) {
-		return envNames, nil
-	}
-
-	envDirEntries, err := os.ReadDir(envsDir)
-	if err != nil {
-		return envNames, errors.WithStackTrace(err)
-	}
-
-	for _, envDirEntry := range envDirEntries {
-		if envDirEntry.IsDir() {
-			envNames = append(envNames, envDirEntry.Name())
-		}
-	}
-
-	return envNames, nil
-}
-
-func createEnv(outputDir string, envName *string, dependencyGraph *graph.TerraformGraph, terragruntOptions *options.TerragruntOptions) error {
-	outPath := outputDir
-	if envName != nil {
-		outPath = filepath.Join(outputDir, *envName)
-	}
-
-	terragruntOptions.Logger.Debugf("Creating env folder: %s", outPath)
-	if err := util.EnsureDirectory(outPath); err != nil {
-		return err
-	}
-
-	parsedTerraformFiles, err := parseAllTerraformFilesInDir(terragruntOptions.WorkingDir)
-	if err != nil {
-		return err
-	}
-
-	moduleNames, err := extractModuleNames(parsedTerraformFiles)
-	if err != nil {
-		return err
-	}
-
-	terragruntOptions.Logger.Debugf("Found the following modules in %s: %v", terragruntOptions.WorkingDir, moduleNames)
-	for _, moduleName := range moduleNames {
-		if err := createModule(moduleName, util.RemoveElementFromList(moduleNames, moduleName), outPath, envName, dependencyGraph, terragruntOptions); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// TerraformFiles is a map from file path to the parsed HCL
-type TerraformFiles map[string]*hclwrite.File
-
-func parseAllTerraformFilesInDir(dir string) (TerraformFiles, error) {
-	out := map[string]*hclwrite.File{}
-
-	terraformFiles, err := filepath.Glob(filepath.Join(dir, "*.tf"))
-	if err != nil {
-		return out, errors.WithStackTrace(err)
-	}
-
-	for _, terraformFile := range terraformFiles {
-		bytes, err := os.ReadFile(terraformFile)
-		if err != nil {
-			return out, errors.WithStackTrace(err)
-		}
-
-		parsedFile, diags := hclwrite.ParseConfig(bytes, terraformFile, hcl.InitialPos)
-		if diags.HasErrors() {
-			return out, errors.WithStackTrace(diags)
-		}
-
-		out[terraformFile] = parsedFile
-	}
-
-	return out, nil
-}
-
-func extractModuleNames(parsedTerraformFiles TerraformFiles) ([]string, error) {
-	moduleNames := []string{}
-
-	for path, parsedFile := range parsedTerraformFiles {
-		for _, block := range parsedFile.Body().Blocks() {
-			if block.Type() == "module" {
-				if len(block.Labels()) != 1 {
-					return moduleNames, fmt.Errorf("Found an invalid module block in %s with more than 1 label: %v", path, block.Labels())
-				}
-
-				moduleNames = append(moduleNames, block.Labels()[0])
-			}
-		}
-	}
-
-	return moduleNames, nil
-}
-
-func createModule(currentModuleName string, otherModuleNames []string, outPath string, envName *string, dependencyGraph *graph.TerraformGraph, terragruntOptions *options.TerragruntOptions) error {
-	modulePath := filepath.Join(outPath, currentModuleName)
-	terragruntOptions.Logger.Debugf("Creating module: %s", modulePath)
-
-	if err := copyOriginalModuleFromWorkingDir(modulePath, terragruntOptions); err != nil {
-		return err
-	}
-
-	if envName != nil {
-		if err := copyEnvContents(modulePath, *envName, terragruntOptions); err != nil {
-			return err
-		}
-	}
-
-	// Since we copied all the contents over, we parse them again, and will modify the copies
-	parsedTerraformFiles, err := parseAllTerraformFilesInDir(modulePath)
-	if err != nil {
-		return err
-	}
-
-	// We are going to modify the graph for each module, so clone it so we aren't modifying the original
-	dependencyGraphClone := dependencyGraph.Clone()
-
-	if err := processFiles(parsedTerraformFiles, currentModuleName, otherModuleNames, envName, dependencyGraphClone, terragruntOptions); err != nil {
-		return err
-	}
-
-	for path, parsedFile := range parsedTerraformFiles {
-		fileContents := parsedFile.Bytes()
-		formattedFileContents := hclwrite.Format(fileContents)
-
-		terragruntOptions.Logger.Debugf("Writing updated contents to %s", path)
-		if err := util.WriteFileWithSamePermissions(path, path, formattedFileContents); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func buildDependencyGraph(terragruntOptions *options.TerragruntOptions) (*graph.TerraformGraph, error) {
-	return graph.GetParsedTerraformGraph(terragruntOptions.WorkingDir)
-}
-
-func copyOriginalModuleFromWorkingDir(modulePath string, terragruntOptions *options.TerragruntOptions) error {
-	terragruntOptions.Logger.Debugf("Copying contents from %s to %s", terragruntOptions.WorkingDir, modulePath)
-	return util.CopyFolderContentsWithFilterImpl(terragruntOptions.WorkingDir, modulePath, nil, false, preprocessorFileCopyFilter)
-}
-
-func copyEnvContents(modulePath string, envName string, terragruntOptions *options.TerragruntOptions) error {
-	envPath := filepath.Join(terragruntOptions.WorkingDir, envsDirName, envName)
-	terragruntOptions.Logger.Debugf("Copying contents from %s to %s", envPath, modulePath)
-	return util.CopyFolderContentsWithFilterImpl(envPath, modulePath, nil, false, preprocessorFileCopyFilter)
-}
-
-// preprocessorFileCopyFilter is a filter that can be used with util.CopyFolderContentsWithFilter to exclude hidden
-// files & folders and state files
-func preprocessorFileCopyFilter(absolutePath string) bool {
-	return !util.TerragruntExcludes(absolutePath) && !strings.HasSuffix(absolutePath, ".tfstate") && !strings.HasSuffix(absolutePath, ".tfstate.backup")
-}
 
 func processFiles(parsedTerraformFiles TerraformFiles, currentModuleName string, otherModuleNames []string, envName *string, dependencyGraph *graph.TerraformGraph, terragruntOptions *options.TerragruntOptions) error {
 	allBlocks := getAllBlocks(parsedTerraformFiles)
@@ -328,65 +114,6 @@ func updateTerraformConfig(terraformBlocks []BlockAndFile, currentModuleName str
 	}
 
 	return nil, nil
-}
-
-type TerraformBackend struct {
-	backendType   string
-	backendConfig *hclwrite.Body
-}
-
-func NewTerraformBackend(block *hclwrite.Block) (*TerraformBackend, error) {
-	if len(block.Labels()) != 1 {
-		return nil, WrongNumberOfLabels{blockType: block.Type(), expectedLabelCount: 1, actualLabels: block.Labels()}
-	}
-
-	return &TerraformBackend{backendType: block.Labels()[0], backendConfig: block.Body()}, nil
-}
-
-func (backend *TerraformBackend) UpdateConfig(currentModuleName string, envName *string, terragruntOptions *options.TerragruntOptions) error {
-	terragruntOptions.Logger.Debugf("Updating backend config...")
-	// TODO! Implement this method to update the path/key values in the config accordingly.
-	switch backend.backendType {
-	case "local":
-	case "remote":
-	case "azurerm":
-	case "consul":
-	case "cos":
-	case "gcs":
-	case "http":
-	case "kubernetes":
-	case "oss":
-	case "pg":
-	case "s3":
-	}
-
-	return nil
-}
-
-var openBraceToken = &hclwrite.Token{
-	Type:  hclsyntax.TokenOBrace,
-	Bytes: []byte("{"),
-}
-
-var closeBraceToken = &hclwrite.Token{
-	Type:  hclsyntax.TokenCBrace,
-	Bytes: []byte("}"),
-}
-
-func (backend *TerraformBackend) ConfigureDataSource(dataSourceBody *hclwrite.Body) error {
-	dataSourceBody.SetAttributeValue("backend", cty.StringVal(backend.backendType))
-	dataSourceBody.AppendNewline()
-
-	// TODO: this needs to have the key/path/etc value updated accordingly!
-	configTokens := []*hclwrite.Token{}
-	configTokens = append(configTokens, openBraceToken)
-	configTokens = append(configTokens, backend.backendConfig.BuildTokens(nil)...)
-	configTokens = append(configTokens, closeBraceToken)
-
-	dataSourceBody.SetAttributeRaw("config", configTokens)
-	dataSourceBody.AppendNewline()
-
-	return nil
 }
 
 func removeOrReplaceModules(moduleBlocks []BlockAndFile, currentModuleName string, otherModuleNames []string, envName *string, backend *TerraformBackend, dependencyGraph *graph.TerraformGraph, terragruntOptions *options.TerragruntOptions) error {
@@ -627,16 +354,4 @@ func removeUnneededVariables(variableBlocks []BlockAndFile, currentModuleName st
 	}
 
 	return nil
-}
-
-// Custom error types
-
-type WrongNumberOfLabels struct {
-	blockType          string
-	expectedLabelCount int
-	actualLabels       []string
-}
-
-func (err WrongNumberOfLabels) Error() string {
-	return fmt.Sprintf("Expected block of type '%s' to have %d labels, but got %d: %v", err.blockType, err.expectedLabelCount, len(err.actualLabels), err.actualLabels)
 }
